+++ src/applications/discharge-wizard/components/questions/DischargeType.jsx
@@ -1,0 +1,65 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Scroll from 'react-scroll';
+import RadioButtons from '@department-of-veterans-affairs/component-library/RadioButtons';
+
+// Relative Imports
+import { shouldShowQuestion } from '../../helpers';
+import { questionLabels } from '../../constants';
+
+const Element = Scroll.Element;
+
+const DischargeType = ({
+  formValues,
+  handleKeyDown,
+  scrollToLast,
+  updateField,
+}) => {
+  const key = '5_dischargeType';
+
+  if (!formValues) {
+    return null;
+  }
+
+  if (!shouldShowQuestion(key, formValues.questions)) {
+    return null;
+  }
+
+  const label = <h4>Which of the following categories best describes you?</h4>;
+  const options = [
+    { label: questionLabels[key][1], value: '1' },
+    { label: questionLabels[key][2], value: '2' },
+  ];
+  const radioButtonProps = {
+    name: key,
+    label,
+    options,
+    key,
+    onValueChange: v => {
+      if (v.dirty) {
+        updateField(key, v.value);
+      }
+    },
+    onMouseDown: scrollToLast,
+    onKeyDown: handleKeyDown,
+    value: {
+      value: formValues[key],
+    },
+  };
+
+  return (
+    <div>
+      <Element name={key} />
+      <RadioButtons {...radioButtonProps} />
+    </div>
+  );
+};
+
+DischargeType.propTypes = {
+  formValues: PropTypes.object.isRequired,
+  handleKeyDown: PropTypes.func,
+  scrollToLast: PropTypes.func,
+  updateField: PropTypes.func,
+};
+
+export default DischargeType;
