+++ src/applications/discharge-wizard/components/questions/Intention.jsx
@@ -1,0 +1,70 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import Scroll from 'react-scroll';
+import RadioButtons from '@department-of-veterans-affairs/component-library/RadioButtons';
+
+// Relative Imports
+import { shouldShowQuestion } from '../../helpers';
+import { questionLabels } from '../../constants';
+
+const Element = Scroll.Element;
+
+const Intention = ({
+  formValues,
+  handleKeyDown,
+  scrollToLast,
+  updateField,
+}) => {
+  const key = '6_intention';
+
+  if (!formValues) {
+    return null;
+  }
+
+  if (!shouldShowQuestion(key, formValues.questions)) {
+    return null;
+  }
+
+  const label = (
+    <h4>
+      Do you want to change your name, discharge date, or anything written in
+      the “other remarks” section of your DD214?
+    </h4>
+  );
+  const options = [
+    { label: `Yes, ${questionLabels[key][1]}`, value: '1' },
+    { label: `No, ${questionLabels[key][2]}`, value: '2' },
+  ];
+  const radioButtonProps = {
+    name: key,
+    label,
+    options,
+    key,
+    onValueChange: v => {
+      if (v.dirty) {
+        updateField(key, v.value);
+      }
+    },
+    onMouseDown: scrollToLast,
+    onKeyDown: handleKeyDown,
+    value: {
+      value: formValues[key],
+    },
+  };
+
+  return (
+    <div>
+      <Element name={key} />
+      <RadioButtons {...radioButtonProps} />
+    </div>
+  );
+};
+
+Intention.propTypes = {
+  formValues: PropTypes.object.isRequired,
+  handleKeyDown: PropTypes.func,
+  scrollToLast: PropTypes.func,
+  updateField: PropTypes.func,
+};
+
+export default Intention;
